Added option to shuffle segments after creation for better DNN training

+++ pipelines/fetchRefSeq.sh
@@ -1,40 +1,35 @@
 #!/bin/bash
-#$ -swd
-#$ -M adam.rivers@ars.usda.gov
-#$ -m abe
-## Specify a run time of 11 hours (or 39600 seconds)
-#$ -l h_rt=12:00:00
-#$ -l ram.c=120G
-#Fetches and sketches RefSeq.
-#Be sure taxonomy is updated first!
-#To use this script outside of NERSC,
+
+#Written by Brian Bushnell
+#Last updated July 19, 2018
+
+#Modified  by Adam Rivers  on January 30, 2019
+
+#Fetches and renames RefSeq.
+#Be sure the taxonomy server is updated first, or run with local taxonomy data!
+#To use this script outside of NERSC when using local taxonomy data,
 #add "taxpath=/path/to/taxonomy_directory/" to each BBTools command
 
+
 #Ensure necessary executables are in your path
-module load bbtools
+#module load bbtools
 module load pigz
-module load samtools
+module laod bbtools
 
-# Fetch RefSeq
-time wget -nv ftp://ftp.ncbi.nlm.nih.gov/refseq/release/complete/*genomic.fna.gz
+# Set this based on where the taxonomy ws downloaded and processed
+TAXTREE=taxonomy/tree.taxtree.gz
 
-#Concatenate into a single file
-time cat *genomic.fna.gz > all.fa.gz
+#Fetch RefSeq
+#time wget -nv ftp://ftp.ncbi.nlm.nih.gov/refseq/release/complete/*genomic.fna.gz
+wget -q -O - ftp://ftp.ncbi.nlm.nih.gov/refseq/release/complete/*genomic.fna.gz \
+| gi2taxid.sh -Xmx2g in=stdin.fa.gz out=renamed.fa.gz pigz=32 unpigz zl=9 \
+tree=$TAXTREE table=null accession=null server ow
 
-#Optionally, delete the old files
-rm *genomic.fna.gz
-
-#Rename by taxID by looking up gi numbers or accessions
-time gi2taxid.sh -Xmx63g in=all.fa.gz out=renamed.fa.gz tree=auto table=auto accession=auto zl=6
-
-# Remove all.fa.gz
-rm all.fa.gz
-
-#Sort by taxonomy and compress with bgzip for random access by pyfaidx
+#Sort by taxonomy.
 #This makes sketching by taxa use much less memory because sketches can be written to disk as soon as they are finished.
-time sortbyname.sh -Xmx63g in=renamed.fa.gz out=stdout.fa taxa tree=auto gi=ignore fastawrap=255 minlen=60 | pbgzip -c -t 16 -6 > sorted.fa.bfgz
+sortbyname.sh -Xmx500g in=renamed.fa.gz memmult=0.33 out=sorted.fa.gz zl=8 pigz=32 taxa tree=$TAXTREE gi=ignore fastawrap=1023 minlen=60 readbufferlen=2 readbuffers=1
 
-
-#remove all.fa.gz and reanmed.fa.gz
-rm renamed.fa.gz
-rm all.fa.gz
+if [ $? -eq 0 ]
+then
+  rm renamed.fa.gz
+fi
